LPS-188559 Extract the register logic for a specific class and remove the immediate property. It will keep the register of the DLProcessor from the PropsKeys.DL_FILE_ENTRY_PROCESSORS behavior.
