+++ modules/apps/document-library/document-library-service/src/main/java/com/liferay/document/library/internal/util/DLProcessorRegistryImpl.java
@@ -14,27 +14,16 @@ import com.liferay.document.library.kernel.util.DLProcessorThreadLocal;
 import com.liferay.exportimport.kernel.lar.PortletDataContext;
 import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
 import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
-import com.liferay.petra.function.UnsafeConsumer;
-import com.liferay.petra.reflect.ReflectionUtil;
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.repository.model.FileVersion;
-import com.liferay.portal.kernel.util.InstanceFactory;
-import com.liferay.portal.kernel.util.MapUtil;
-import com.liferay.portal.kernel.util.PortalClassLoaderUtil;
-import com.liferay.portal.kernel.util.PropsKeys;
-import com.liferay.portal.kernel.util.PropsUtil;
 import com.liferay.portal.kernel.xml.Element;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
 
 import org.osgi.framework.BundleContext;
-import org.osgi.framework.ServiceRegistration;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Deactivate;
@@ -45,7 +34,7 @@ import org.osgi.service.component.annotations.Modified;
  */
 @Component(
 	configurationPid = "com.liferay.document.library.configuration.DLFileEntryConfiguration",
-	immediate = true, service = DLProcessorRegistry.class
+	service = DLProcessorRegistry.class
 )
 public class DLProcessorRegistryImpl implements DLProcessorRegistry {
 
@@ -182,40 +171,14 @@ public class DLProcessorRegistryImpl implements DLProcessorRegistry {
 		_dlFileEntryConfiguration = ConfigurableUtil.createConfigurable(
 			DLFileEntryConfiguration.class, properties);
 
-		_bundleContext = bundleContext;
-
 		_dlProcessorServiceTrackerMap =
 			ServiceTrackerMapFactory.openSingleValueMap(
 				bundleContext, DLProcessor.class, "type");
-
-		ClassLoader classLoader = PortalClassLoaderUtil.getClassLoader();
-
-		for (String dlProcessorClassName : _DL_FILE_ENTRY_PROCESSORS) {
-			DLProcessor dlProcessor = (DLProcessor)InstanceFactory.newInstance(
-				classLoader, dlProcessorClassName);
-
-			dlProcessor.afterPropertiesSet();
-
-			_register(dlProcessor);
-
-			_dlProcessors.add(dlProcessor);
-		}
 	}
 
 	@Deactivate
 	protected void deactivate() throws Exception {
 		_dlProcessorServiceTrackerMap.close();
-
-		UnsafeConsumer.accept(
-			_dlProcessors,
-			dlProcessor -> {
-				_unregister(dlProcessor);
-
-				dlProcessor.destroy();
-			},
-			Exception.class);
-
-		_dlProcessors.clear();
 	}
 
 	private FileVersion _getLatestFileVersion(
@@ -240,48 +203,11 @@ public class DLProcessorRegistryImpl implements DLProcessorRegistry {
 		}
 	}
 
-	private void _register(DLProcessor dlProcessor) {
-		Class<?>[] classes = ReflectionUtil.getInterfaces(dlProcessor);
-
-		String[] classNames = new String[classes.length];
-
-		for (int i = 0; i < classes.length; i++) {
-			classNames[i] = classes[i].getName();
-		}
-
-		ServiceRegistration<?> serviceRegistration =
-			_bundleContext.registerService(
-				classNames, dlProcessor,
-				MapUtil.singletonDictionary("type", dlProcessor.getType()));
-
-		ServiceRegistration<?> previousServiceRegistration =
-			_serviceRegistrations.put(dlProcessor, serviceRegistration);
-
-		if (previousServiceRegistration != null) {
-			previousServiceRegistration.unregister();
-		}
-	}
-
-	private void _unregister(DLProcessor dlProcessor) {
-		ServiceRegistration<?> serviceRegistration =
-			_serviceRegistrations.remove(dlProcessor);
-
-		serviceRegistration.unregister();
-	}
-
-	private static final String[] _DL_FILE_ENTRY_PROCESSORS =
-		PropsUtil.getArray(PropsKeys.DL_FILE_ENTRY_PROCESSORS);
-
 	private static final Log _log = LogFactoryUtil.getLog(
 		DLProcessorRegistryImpl.class);
 
-	private BundleContext _bundleContext;
 	private volatile DLFileEntryConfiguration _dlFileEntryConfiguration;
-	private final List<DLProcessor> _dlProcessors = new ArrayList<>(
-		_DL_FILE_ENTRY_PROCESSORS.length);
 	private ServiceTrackerMap<String, DLProcessor>
 		_dlProcessorServiceTrackerMap;
-	private final Map<DLProcessor, ServiceRegistration<?>>
-		_serviceRegistrations = new ConcurrentHashMap<>();
 
 }
